+++ src/Areas/KeyVault/Commands/KeyVaultJsonContext.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System.Text.Json.Serialization;
+using AzureMcp.Areas.KeyVault.Commands.Certificate;
 using AzureMcp.Areas.KeyVault.Commands.Key;
 using AzureMcp.Areas.KeyVault.Commands.Secret;
 
@@ -10,7 +11,12 @@ namespace AzureMcp.Commands.KeyVault;
 [JsonSerializable(typeof(KeyListCommand.KeyListCommandResult))]
 [JsonSerializable(typeof(KeyGetCommand.KeyGetCommandResult))]
 [JsonSerializable(typeof(KeyCreateCommand.KeyCreateCommandResult))]
+[JsonSerializable(typeof(SecretListCommand.SecretListCommandResult))]
+[JsonSerializable(typeof(SecretCreateCommand.SecretCreateCommandResult))]
 [JsonSerializable(typeof(SecretGetCommand.SecretGetCommandResult))]
+[JsonSerializable(typeof(CertificateListCommand.CertificateListCommandResult))]
+[JsonSerializable(typeof(CertificateGetCommand.CertificateGetCommandResult))]
+[JsonSerializable(typeof(CertificateCreateCommand.CertificateCreateCommandResult))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
 internal sealed partial class KeyVaultJsonContext : JsonSerializerContext
 {